# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.CreateWorkoutExercisePayload:
    properties:
      duration:
        type: integer
      exercise_id:
        type: integer
    required:
    - duration
    - exercise_id
    type: object
  handlers.CreateWorkoutPayload:
    properties:
      description:
        type: string
      exercises:
        items:
          $ref: '#/definitions/handlers.CreateWorkoutExercisePayload'
        type: array
      name:
        type: string
      tutorial_link:
        type: string
    required:
    - exercises
    - name
    type: object
...
  handlers.WorkoutsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/store.Workout'
        type: array
      total_count:
        type: integer
    type: object
...
  store.Exercise:
    properties:
      created_at:
        type: string
      description:
        type: string
      duration:
        type: integer
      id:
        type: integer
      is_duration:
        type: boolean
      like:
        type: integer
      muscles:
        items:
          type: string
        type: array
      name:
        type: string
      tutorial_link:
        type: string
      user_id:
        type: integer
    type: object
...
  store.Workout:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      likes:
        type: integer
      name:
        type: string
      rating:
        type: number
      reviews_count:
        type: integer
      tutorial_link:
        type: string
      user_id:
        type: integer
      user_liked:
        type: boolean
      workout_exercises:
        items:
          $ref: '#/definitions/store.WorkoutExercises'
        type: array
    type: object
  store.WorkoutExercises:
    properties:
      duration:
        type: integer
      exercise:
        $ref: '#/definitions/store.Exercise'
      exercise_id:
        type: integer
      workout_id:
        type: integer
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for Atom Fit API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Atom Fit API
paths:
...
  /workouts:
    post:
      consumes:
      - application/json
      description: Create a new workout
      parameters:
      - description: Create Workout Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateWorkoutPayload'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - ApiKeyAuth: []
      summary: Create a new workout
      tags:
      - workouts
  /workouts/:
    get:
      consumes:
      - application/json
      description: Get all workouts
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: Sort
        in: query
        name: sort
        type: string
      - description: Tags
        in: query
        name: tags
        type: string
      - description: Search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.WorkoutsResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all workouts
      tags:
      - workouts
  /workouts/{workoutID}:
    delete:
      consumes:
      - application/json
      description: Delete workout by ID
      parameters:
      - description: Workout ID
        in: path
        name: workoutID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - ApiKeyAuth: []
      summary: Delete workout by ID
      tags:
      - workouts
    get:
      consumes:
      - application/json
      description: Get workout by ID
      parameters:
      - description: Workout ID
        in: path
        name: workoutID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.Workout'
      summary: Get workout by ID
      tags:
      - workouts
    patch:
      consumes:
      - application/json
      description: Update workout by ID
      parameters:
      - description: Workout ID
        in: path
        name: workoutID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - ApiKeyAuth: []
      summary: Update workout by ID
      tags:
      - workouts
...
  /workouts/user/{userID}:
    get:
      consumes:
      - application/json
      description: Get user workouts
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: Sort
        in: query
        name: sort
        type: string
      - description: userID
        in: path
        name: userID
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.Workout'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Get user workouts
      tags:
      - workouts
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"